update solid-register to 0.1.3
