+++ packages/active-element/package.json
@@ -49,7 +49,7 @@
   ],
   "devDependencies": {
     "prettier": "^2.5.0",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "tslib": "^2.3.1",
     "tsup": "^5.10.0",
     "unocss": "0.21.2",
