+++ packages/clipboard/package.json
@@ -52,7 +52,7 @@
   "devDependencies": {
     "jsdom": "18.1.1",
     "prettier": "^2.5.1",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "tslib": "^2.3.1",
     "tsup": "^5.11.1",
     "typescript": "^4.5.3",
