+++ packages/composites/src/createCompositeComputed.ts
@@ -1,5 +1,5 @@
 import { createComputed, createRoot, onCleanup } from "solid-js";
-import type { StopEffect, WatchOptions, ModifierReturn, EffectCallback } from ".";
+import type { StopEffect, WatchOptions, ModifierReturn, EffectCallback } from "./types";
 import { createComputationWatcher, Fn, parseCompositeArgs } from "./common";
 
 /**
