+++ packages/composites/src/createCompositeRenderEffect.ts
@@ -1,5 +1,5 @@
 import { createRenderEffect, createRoot, onCleanup } from "solid-js";
-import type { StopEffect, WatchOptions, ModifierReturn, EffectCallback } from ".";
+import type { StopEffect, WatchOptions, ModifierReturn, EffectCallback } from "./types";
 import { createComputationWatcher, Fn, parseCompositeArgs } from "./common";
 
 /**
