+++ packages/devices/package.json
@@ -53,7 +53,7 @@
     "jsdom": "^18.1.1",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "tslib": "2.3.1",
     "tsup": "^5.10.1",
     "typescript": "^4.5.2",
