+++ packages/event-bus/src/eventHub.ts
@@ -1,4 +1,6 @@
-import { ClearListeners, Unsubscribe, createEventBus, createEmitter, GenericListener } from ".";
+import { createEventBus } from './eventBus';
+import { createEmitter } from './emitter';
+import { ClearListeners, Unsubscribe, GenericListener } from "./types";
 import { Accessor } from "solid-js";
 import { Keys, Values } from "@solid-primitives/utils";
