+++ packages/event-props/package.json
@@ -43,7 +43,7 @@
   "devDependencies": {
     "jsdom": "^18.1.1",
     "prettier": "^2.5.0",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "tslib": "^2.3.1",
     "tsup": "^5.11.1",
     "typescript": "^4.5.2",
