+++ packages/mouse/src/mousePosition.ts
@@ -1,7 +1,6 @@
 import { ClearListeners } from "@solid-primitives/event-listener";
 import { isClient, createCallbackStack } from "@solid-primitives/utils";
 import { Accessor, batch, createSignal } from "solid-js";
-import { Position } from ".";
 import { addListener } from "./common";
 
 export interface MouseOptions {
@@ -86,3 +85,43 @@ export function createMousePosition(options: MouseOptions = {}): [
     toCleanup.execute
   ];
 }
+
+export interface Position {
+  x: number;
+  y: number;
+}
+export interface RelativeToElementValues extends Position {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+}
+
+/**
+ * Turn position relative to the page, into position relative to an element.
+ */
+export const posRelativeToElement = (
+  pageX: number,
+  pageY: number,
+  el: Element
+): RelativeToElementValues => {
+  const bounds = el.getBoundingClientRect(),
+    top = bounds.top + window.screenY,
+    left = bounds.left + window.scrollX;
+  return {
+    x: pageX - left,
+    y: pageY - top,
+    top,
+    left,
+    width: bounds.width,
+    height: bounds.height
+  };
+};
+
+/**
+ * Turn position relative to the page, into position relative to the screen.
+ */
+export const posRelativeToScreen = (pageX: number, pageY: number): Position => ({
+  x: pageX - window.scrollX,
+  y: pageY - window.screenY
+});
