+++ packages/props/package.json
@@ -41,7 +41,7 @@
     "esbuild-plugin-solid": "^0.4.1",
     "jsdom": "18.1.1",
     "prettier": "^2.0.5",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "solid-testing-library": "^0.2.0",
     "tslib": "^2.3.1",
     "tsup": "^5.10.1",
