+++ packages/script-loader/package.json
@@ -49,7 +49,7 @@
   "devDependencies": {
     "jsdom": "18.1.1",
     "prettier": "^2.0.5",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "solid-testing-library": "^0.2.0",
     "tslib": "^2.3.1",
     "tsup": "^5.10.1",
