+++ packages/storage/package.json
@@ -45,7 +45,7 @@
   "devDependencies": {
     "jsdom": "18.1.1",
     "prettier": "^2.0.5",
-    "solid-register": "0.1.1",
+    "solid-register": "0.1.3",
     "tsup": "^5.10.1",
     "typescript": "^4.5.2",
     "uvu": "^0.5.2"
